requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
requests>=2.26.0
python-dotenv==1.0.0
pytest>=6.2.5
flask==2.3.3
flask-cors==4.0.0
requests
beautifulsoup4==4.12.2
aiohttp==3.8.5
cachetools==5.3.1
certifi==2023.7.22